[AI] SISC2-36 [FEAT] 최종 리포트 저장 구현
