+++ AI/finder/__init__.py
@@ -1,0 +1,3 @@
+# AI/finder/__init__.py
+from .main import run_finder
+__all__ = ["run_finder"]
