+++ AI/transformer/__init__.py
@@ -1,0 +1,3 @@
+# AI/finder/__init__.py
+from .main import run_transformer
+__all__ = ["run_transformer"]
